Fixes an overflow issue when exactly 256 bones are used (#617)

+++ filament/src/components/RenderableManager.cpp
@@ -286,7 +286,7 @@ void FRenderableManager::create(
                     driver.createUniformBuffer(count * sizeof(PerRenderableUibBone),
                             driver::BufferUsage::DYNAMIC),
                     UniformBuffer{ count * sizeof(PerRenderableUibBone) },
-                    (uint8_t)count
+                    count
             });
             assert(bones);
             if (bones) {
